server/requirements.txt
aiohttp>=3.9.0
opencv-python>=4.8.0
pyngrok>=7.0.0
python-dotenv>=1.0.0
